+++ cmd/frontend/graphqlbackend/search_repositories_test.go
@@ -3,8 +3,11 @@ package graphqlbackend
 import (
 	"context"
 	"errors"
+	"sort"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	searchbackend "github.com/sourcegraph/sourcegraph/internal/search/backend"
@@ -44,80 +47,73 @@ func TestSearchRepositories(t *testing.T) {
 		}
 	}
 
-	t.Run("search for all repositories", func(t *testing.T) {
-		q, err := query.ParseAndCheck("type:repo")
-		if err != nil {
-			t.Fatal(err)
-		}
-		args := search.TextParameters{
-			PatternInfo: &search.TextPatternInfo{Pattern: "", IsRegExp: true, FileMatchLimit: 1, PathPatternsAreRegExps: true, PathPatternsAreCaseSensitive: false, PatternMatchesContent: true, PatternMatchesPath: true},
-			Repos:       repositories,
-			Query:       q,
-			Zoekt:       zoekt,
-		}
-		res, _, err := searchRepositories(context.Background(), &args, int32(100))
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(res) != len(repositories) {
-			t.Errorf("expected all repository results, but got %v", len(res))
-		}
+	cases := []struct {
+		name string
+		q    string
+		want []string
+	}{{
+		name: "all",
+		q:    "type:repo",
+		want: []string{"bar/one", "foo/no-match", "foo/one"},
+	}, {
+		name: "pattern filter",
+		q:    "type:repo foo/one",
+		want: []string{"foo/one"},
+	}, {
+		name: "repohasfile",
+		q:    "foo type:repo repohasfile:f.go",
+		want: []string{"foo/one"},
+	}, {
+		name: "case yes match",
+		q:    "foo case:yes",
+		want: []string{"foo/no-match", "foo/one"},
+	}, {
+		name: "case no match",
+		q:    "Foo case:no",
+		want: []string{"foo/no-match", "foo/one"},
+	}, {
+		name: "case exclude all",
+		q:    "Foo case:yes",
+		want: []string{},
+	}}
 
-	})
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			q, err := query.ParseAndCheck(tc.q)
+			if err != nil {
+				t.Fatal(err)
+			}
 
-	t.Run("search for all repositories where the repo name includes 'foo/one'", func(t *testing.T) {
-		q, err := query.ParseAndCheck("type:repo foo/one")
-		if err != nil {
-			t.Fatal(err)
-		}
-		args := search.TextParameters{
-			PatternInfo: &search.TextPatternInfo{Pattern: "foo/one", IsRegExp: true, FileMatchLimit: 1, PathPatternsAreRegExps: true, PathPatternsAreCaseSensitive: false, PatternMatchesContent: true, PatternMatchesPath: true},
-			Repos:       repositories,
-			Query:       q,
-			Zoekt:       zoekt,
-		}
-		res, _, err := searchRepositories(context.Background(), &args, int32(100))
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(res) != 1 {
-			t.Errorf("expected only one repository result `foo/one`, but got %v", len(res))
-		}
-		r, ok := res[0].ToRepository()
-		if !ok {
-			t.Fatalf("expected repo result")
-		}
-		if r.repo.Name != "foo/one" {
-			t.Errorf("expected the repository result to be `foo/one`, but got another repo")
-		}
-	})
+			pattern, err := getPatternInfo(q, &getPatternInfoOptions{fileMatchLimit: 1})
+			if err != nil {
+				t.Fatal(err)
+			}
 
-	t.Run("search for all repositories where the repo name includes 'foo' and the repo has a file path matching 'f.go'", func(t *testing.T) {
-		q, err := query.ParseAndCheck("foo type:repo repohasfile:f.go")
-		if err != nil {
-			t.Fatal(err)
-		}
-		args := search.TextParameters{
-			PatternInfo: &search.TextPatternInfo{Pattern: "foo", IsRegExp: true, FileMatchLimit: 1, FilePatternsReposMustInclude: []string{"f.go"}, PathPatternsAreRegExps: true, PathPatternsAreCaseSensitive: false, PatternMatchesContent: true, PatternMatchesPath: true},
-			Repos:       repositories,
-			Query:       q,
-			Zoekt:       zoekt,
-		}
-		res, _, err := searchRepositories(context.Background(), &args, int32(100))
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(res) != 1 {
-			t.Errorf("expected only one repository result `foo/one`, but got %v", len(res))
-		}
-		r, ok := res[0].ToRepository()
-		if !ok {
-			t.Fatalf("expected repo result")
-		}
-		if r.repo.Name != "foo/one" {
-			t.Errorf("expected the repository result to be `foo/one`, but got another repo")
-		}
-	})
+			results, _, err := searchRepositories(context.Background(), &search.TextParameters{
+				PatternInfo: pattern,
+				Repos:       repositories,
+				Query:       q,
+				Zoekt:       zoekt,
+			}, int32(100))
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			var got []string
+			for _, res := range results {
+				r, ok := res.ToRepository()
+				if !ok {
+					t.Fatal("expected repo result")
+				}
+				got = append(got, string(r.repo.Name))
+			}
+			sort.Strings(got)
+
+			if !cmp.Equal(tc.want, got, cmpopts.EquateEmpty()) {
+				t.Errorf("mismatch (-want +got):\n%s", cmp.Diff(tc.want, got))
+			}
+		})
+	}
 }
 
 func TestRepoShouldBeAdded(t *testing.T) {