+++ cmd/frontend/internal/app/sign_out.go
@@ -58,7 +58,7 @@ func renderSignoutPageTemplate(w http.ResponseWriter, r *http.Request, signoutUR
 		return
 	}
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	buf.WriteTo(w)
+	_, _ = buf.WriteTo(w)
 }
 
 var signoutPageTemplate = template.Must(template.New("").Parse(`