+++ web/src/search/input/RegexpToggle.tsx
@@ -7,12 +7,14 @@ import { Subscription, fromEvent } from 'rxjs'
 import { filter } from 'rxjs/operators'
 import { PatternTypeProps, CaseSensitivityProps } from '..'
 import { FiltersToTypeAndValue } from '../../../../shared/src/search/interactive/util'
+import classNames from 'classnames'
 
-interface RegexpToggleProps extends PatternTypeProps, CaseSensitivityProps {
+export interface RegexpToggleProps extends PatternTypeProps, CaseSensitivityProps {
     navbarSearchQuery: string
     history: H.History
     filtersInQuery?: FiltersToTypeAndValue
     hasGlobalQueryBehavior?: boolean
+    className?: string
 }
 
 export class RegexpToggle extends React.Component<RegexpToggleProps> {
@@ -46,7 +48,10 @@ export class RegexpToggle extends React.Component<RegexpToggleProps> {
             <div
                 ref={this.toggleCheckbox}
                 onClick={this.toggle}
-                className="btn btn-icon icon-inline query-input2__toggle e2e-regexp-toggle"
+                className={classNames(
+                    'btn btn-icon icon-inline query-input2__toggle e2e-regexp-toggle',
+                    this.props.className
+                )}
                 role="checkbox"
                 aria-checked={isRegexp}
                 aria-label="Regular expression toggle"